Introduced freeform canvas to draw freeforms
