+++ src/apps/demo_sprites/demo_sprites_meta_data.rb
@@ -25,7 +25,7 @@ class DemoSpritesMetaData
   end
 
   def self.canvas
-    GridCanvas
+    FreeformCanvas
   end
 
   def self.render_attributes
